+++ client/src/components/pages/CraigForms.js
@@ -896,6 +896,10 @@ function craigForms(craig) {
           archive: craig.logdna.archive,
           platform_logs: craig.logdna.platform_logs,
         },
+        {
+          store_secrets: craig.logdna.store_secrets,
+          secrets_manager: craig.logdna.secrets_manager,
+        },
       ],
     },
     object_storage: {
@@ -1387,6 +1391,10 @@ function craigForms(craig) {
         {
           plan: craig.sysdig.plan,
           platform_logs: craig.sysdig.platform_logs,
+        },
+        {
+          store_secrets: craig.sysdig.store_secrets,
+          secrets_manager: craig.sysdig.secrets_manager,
         },
       ],
     },