+++ CHANGELOG.md
@@ -13,10 +13,12 @@ All notable changes to this project will be documented in this file.
 ### Upgrade Notes
 
 - When creating a GRE Tunnel a Remote Gateway IP and Local Gateway IP must be entered in the GRE Tunnel Form to allow for Terraform to be generated successfully
+- Power VS SSH Keys are now optional for Power VS Instances and FalconStor VTL
 
 ### Features
 
 - Added documentation for Power VS Shared Processor Pools and Power VS Placement Groups
+- LogDNA and Sysdig keys can now be stored for later access in a Secrets Manager instance
 
 ## 1.16.0